Minor fixes and improvements in existing tests
* Fix old API names
* Fix LB cast to double
* Increase default loads and parallel granularity
* Improve MPI+OpenMP test:
  - Use performance libs instead of debug
  - Set some OMP env. vars to improve performance in DLB
  - Document the MPI no-binding flag

+++ doc/examples/statistics/mpi_ompss_pils.c
@@ -58,7 +58,7 @@ double LoadUnbalance(int loads[], int how_many) {
         sum+=loads[i];
         max=MAX(max, loads[i]);
     }
-    return sum / (how_many*max);
+    return (double)sum / (how_many*max);
 }
 
 #pragma omp task
